Fixed inconsistency between errors
When u/dvoting a claim, if you passed in a non-int number into the second
argument, it would give the error "times not int". When u/dvoting a
comment, if you passed in a non-int number into the second argument, it
would give the error "times not an int". The comment u/dvoting error
message was changed to be equal to the claim error.